get unit tests working after dir restructure
