+++ package.json
@@ -24,7 +24,7 @@
     "codeclimate": "npm run coverage -- --reporter=lcov --reporter=text-lcov | ./node_modules/codeclimate-test-reporter/bin/codeclimate.js",
     "lint": "eslint . ",
     "test-cli": "NODE_ENV=production NODE_PATH=. _mocha --require ./tests/spec/specSetup.js --opts ./tests/config/mocha.opts 'bin/**/*.spec.js'",
-    "test-unit": "NODE_ENV=production NODE_PATH=. _mocha --require ./tests/spec/specSetup.js --opts ./tests/config/mocha.opts '{tests/spec,tests/e2e,components}/**/*.spec.js'",
+    "test-unit": "NODE_ENV=production NODE_PATH=. _mocha --require ./tests/spec/specSetup.js --opts ./tests/config/mocha.opts '{src,tests/e2e,components}/**/*.spec.js'",
     "test-integration": "_mocha -t 7000 --opts ./tests/config/mocha.opts 'tests/integration/**/*.spec.js' --config ./app/config/application.js --config './app/config/integrationOverrides.json'",
     "pretest-e2e": "npm run link:e2e-tests",
     "test-e2e": "npm run nightwatch -- --components=test-components",
