+++ src/index.js
@@ -3,7 +3,7 @@ const devConfig = require('./config.dev');
 const prodConfig = require('./config');
 
 module.exports = function bootstrap(initialState) {
-  console.log(`process.env.NODE_ENV`, process.env.NODE_ENV)
+
   const breadboardOptions = process.env.NODE_ENV === 'production'
     ? prodConfig(initialState)
     : devConfig(initialState);
