+++ tests/commonMocks.js
