+++ tests/e2e/components/test-redux/components/Price/Price.spec.js
@@ -1,4 +1,4 @@
-import helper from 'tests/helper';
+import helper from '../../../../../../tests/helper';
 import React from 'react';
 import {shallow, mount} from 'enzyme';
 import Price from './Price';
