+++ tests/e2e/components/test-redux/components/ProductListItem/ProductListItem.spec.js
@@ -1,4 +1,4 @@
-import helper from 'tests/helper';
+import helper from '../../../../../../tests/helper';
 import React from 'react';
 import {shallow} from 'enzyme';
 import Product from './ProductListItem';
